% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_grade_weight.R
\name{calc_grade_weight}
\alias{calc_grade_weight}
\title{Calculate weight for usabilty grade}
\usage{
calc_grade_weight(grade, max_grade, min_grade, min_weight)
}
\arguments{
\item{grade}{Input value to be scaled.}

\item{max_grade}{Number corresponding to best possible grade.}

\item{min_grade}{Number corresponding to worst possible grade.}

\item{min_weight}{Number between 0 and 1 corresponding to weight chosen for worst possible grade.}
}
\value{
Number between 0 and 1 corresponding calculated weight.
}
\description{
Scales usability grade of any format to between 0 and 1 for use as weight.
}
\examples{
calc_grade_weight(grade=2, max_grade=0, min_grade=5, min_weight=0.5)
df <- data.frame(Grade=c(1, 0, 2, 1, 1, 2, 2, 2, 0))
df$GradeWeights <- sapply(
 1:nrow(df), 
 function(i) {
   calc_grade_weight(
     grade=df$Grade[i], 
     max_grade=2,
     min_grade=0,
     min_weight=0.6
   )
 }
)
}
