% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metric.R
\name{calc_metric}
\alias{calc_metric}
\title{Calculate metric value}
\usage{
calc_metric(fundamental_vars = c(), additional_vars = c(), betas = c())
}
\arguments{
\item{fundamental_vars}{Vector of numbers between 0 and 1.}

\item{additional_vars}{Vector of numbers between 0 and 1.}

\item{betas}{Vector of numbers between 0 and 1 that sum to 1.}
}
\value{
A number between 0 and 1.
}
\description{
Calculates a number between 0 and 1, using given inputs, corresponding to the metric value
}
\examples{
calc_metric(c(0.2, 0.3), c(1, 0), c(0.5, 0.3, 0.2))
}
