% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_uncertainty_weight.R
\name{calc_uncertainty_weight}
\alias{calc_uncertainty_weight}
\title{Calculate weight as estimate of uncertainty}
\usage{
calc_uncertainty_weight(scaled_sample_size, grade_weight)
}
\arguments{
\item{scaled_sample_size}{Single numeric or vector of numeric input value/s between 0 and 1.}

\item{grade_weight}{Single numeric or vector of numeric input value/s between 0 and 1.}
}
\value{
Number or vector of numbers between 0 and 1 corresponding calculated weights.
}
\description{
Combines grade weight with scaled sample size to generate overall weight.
}
\examples{
calc_uncertainty_weight(scaled_sample_size = 0.7, grade_weight = 0.8)
calc_uncertainty_weight(
   scaled_sample_size = c(0.7, 0.5, 0.3, 0.8), 
   grade_weight = c(0.8, 0.2, 0.4, 0.4)
)
}
