% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roam-package.R
\docType{package}
\name{roam-package}
\alias{roam-package}
\title{roam: Remote Objects with Active-Binding Magic}
\description{
Provide helper functions for package developers to create active bindings that looks like data embedded in the package, but are downloaded from remote sources.
}
\author{
\strong{Maintainer}: Yangzhuoran Fin Yang \email{yangyangzhuoran@gmail.com} (\href{https://orcid.org/0000-0002-1232-8017}{ORCID})

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}
}

}
\keyword{internal}
