% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sample_size.R
\name{scale_sample_size}
\alias{scale_sample_size}
\title{Scale sample size for use as weights}
\usage{
scale_sample_size(x, method = "sigmoid")
}
\arguments{
\item{x}{Vector of numeric input values.}

\item{method}{Character string either "sigmoid" (default) or "log-linear".}
}
\value{
Vector of numbers between 0 and 1 corresponding calculated weights.
}
\description{
Scales continuous input to between 0 and 1 for use as weights.
}
\examples{
scale_sample_size(c(100, 50, 150), method="log-linear")
df <- data.frame(SampleSize=c(10, 15, 100, 30, 50, 90, 140))
scale_sample_size(df$SampleSize)
}
