% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAM.R
\name{OAM}
\alias{OAM}
\title{Compute the Opening Angle Method}
\usage{
OAM(
  watermask,
  precision = 360,
  save_im = T,
  fn_r = NULL,
  ext = NULL,
  no_cores = 3,
  fork = F,
  socket = F,
  native = F
)
}
\arguments{
\item{watermask}{Numeric: binary matrix where 1 indicates water and 0 indicates land. Should be passed with georeferenced info.}

\item{precision}{Integer: Number of discrete angles to split circle into (angles tested given by 360/precision)}

\item{save_im}{Logical: Whether or not the output should be written to a geoTIFF via stars}

\item{fn_r}{Character: Output file name if save_im is true.}

\item{ext}{st_crs: Geographic extent information from the stars object.}

\item{no_cores}{Integer: Number of cores to use if running in parallel. Parallel is assumed by default.}

\item{fork}{Logical: whether to use fork-type parallelization. See details. Only one of fork, socket, and native should be true.}

\item{socket}{Logical: whether to use socket-type parallelization.}

\item{native}{Logical: whether to use native-type parallelization.}
}
\value{
Numeric: matrix with same dimensions as watermask
}
\description{
test 1
}
\details{
This function computes the opening angle method based on Shaw et al., 2008 and through an a modification approximation used in Liang et al., (2015a). This approximation uses the rayshader package to efficiently cast rays from an angle phi to every pixel in the query set Q.
}
\examples{
# See example in Github repo.
}
