% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAM.R
\name{OAM}
\alias{OAM}
\title{Opening Angle Method}
\usage{
OAM(
  watermask,
  precision = 360,
  save_im = T,
  fn_r = NULL,
  geoInfo = NULL,
  no_cores = 3,
  parallel = 0
)
}
\arguments{
\item{watermask}{Numeric or stars: binary n x m matrix where 1 indicates water and 0 indicates land. The input image should ideally be passed as a georeferenced stars object.}

\item{precision}{Integer: Number (\code{p}) of discrete sectors to split circle into (angles tested given by 360/precision)}

\item{save_im}{Logical: Whether or not the output should be written to a geoTIFF via `stars``}

\item{fn_r}{Character: Output file name if save_im is true.}

\item{geoInfo}{st_dimensions: Georaphic information including coordinate reference system, resolution, etc. of the georeferenced output. If a stars object is passed to \code{map}, its st_dimensions properties will be used.}

\item{no_cores}{Integer: Number of cores to use if running in parallel. No parallelization is assumed by default.}

\item{parallel}{Integer: Parallelization scheme (see details)}
}
\value{
Numeric: matrix with same dimensions as watermask
}
\description{
Compute the opening angle method
}
\details{
This function implements the Opening Angle Method. This R based version uses an approximation of the OAM proposed by John Shaw and commonly used by the delta morphology community. See the \code{ROAM} package vignette for details.

The parallel parameter takes integer values from 0 to 3.  0 indicates no parallelization. Three parallelization schemes are possible: 1 use rayshaders-native parallelization which dissects an input scene into chunks before raytracing for a given angle; 2 and 3 split up the computation of the raytracing into \code{no_cores} blocks and then tallies the results. \code{2} indicates the use socket-based parallelization, necessary on windows machines; 3 use fork-based parallelization, possible on *nix machines.
}
\examples{
\dontrun{
watermask = stars::read_stars("SaoFrancisco.tif")
plot(watermask, main = "Sao Francisco")
OAM_map = OAM(watermask, precision = 720, save_im = T,
fn_r = "SaoFrancisco_OAM.tif", geoInfo = NULL, no_cores = 3,
parallel = 2)
}
}
