% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roam-package.R
\docType{package}
\name{roam-package}
\alias{roam-package}
\title{roam: Remote Objects with Active-binding Magic}
\description{
Provide helper functions for package developers to create active bindings that looks like data embedded in the package, but are downloaded from remote sources.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Yangzhuoran Fin Yang \email{yangyangzhuoran@gmail.com}
}

}
\keyword{internal}
