% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roam.R
\name{roam}
\alias{roam}
\alias{new_roam}
\alias{roam_update}
\alias{roam_install}
\alias{roam_version}
\alias{roam_delete}
\alias{roam_activate}
\alias{roam_activate_all}
\title{Create and manage roam object and their active bindings}
\usage{
new_roam(package, name, obtainer, ...)

roam_update(x)

roam_install(x, version = NULL)

roam_version(version = NULL, package = NULL, name = NULL)

roam_delete(x)

roam_activate(x)

roam_activate_all(package)
}
\arguments{
\item{package}{the name of the package as a string.}

\item{name}{the name of the roam object.
Should be the same as the name to which the roam object is assigned.}

\item{obtainer}{a package writer/roam object creator defined function to download data/object.
The first argument should be the version number user wants to download.
If input \code{NULL}, the obtainer function should download the latest version.}

\item{...}{optional arguments to \code{obtainer}.}

\item{x}{roam active binding}

\item{version}{In \code{\link[=roam_install]{roam_install()}} version of the data to install. If \code{NULL}, the latest version.
In \code{\link[=roam_version]{roam_version()}}, the version of the currently downloading data.}
}
\value{
\code{new_roam} returns a function with class \code{roam_object}.

\code{roam_update} returns the updated local cache of the roam active binding

\code{roam_install} returns the installed local cache of the roam active binding

\code{roam_version} returns the version.

All the other functions return invisible \code{NULL}.
}
\description{
Helper functions for package developers to create active bindings
that looks like data embedded in the package, but are downloaded from remote sources.
}
\details{
Users of the package using roam data object can treat the roam active bindings
as if they are regular data embedded in the package.
The first time a user calls the roam active binding, they will be prompted to
download the data using the \code{obtainer} function.
The \code{obtainer} function defines how the package developer wants to
download or generate data.
Once the data are downloaded, they will be cached locally using \code{rappdirs}.
The users can then use the data object as normal.

\code{new_roam} creates a roam object using the package writer/roam object creator defined \code{obtainer} function.
The roam object created using \code{new_roam} is not an active binding.
The active bindings are not preserved during package installation,
so the package developer needs to activate the roam object and turn it
into an active binding in the \link{.onLoad} function
using either \code{roam_activate} or \code{roam_activate_all}.

\code{roam_activate} takes one roam object and activates it.
\code{roam_activate_all} looks through the namespace
and activates all the roam objects in the package.

If there are a lot of objects in the package,
calling \code{roam_activate} on each roam object in \link{.onLoad}
might save some package loading time than calling \code{roam_activate_all}
once.
}
\section{Functions}{
\itemize{
\item \code{roam_update()}: Update the local cache of the roam active binding
using the package writer/roam object creator defined obtainer function

\item \code{roam_install()}: Install (Download) the local cache of the roam active binding
of a specific version
using the package writer/roam object creator defined obtainer function

\item \code{roam_version()}: \itemize{
\item Package writer: When used inside the obtainer function, save the currently downloading version number.
\code{package} and \code{name} should not be specified.
\item User: When used outside the obtainer function, find the current version of a roam object in a package.
\code{version} should not be specified.
}

\item \code{roam_delete()}: Delete the local cache of the roam active binding

\item \code{roam_activate()}: Activate a roam object to an active binding.
Used in the \link{.onLoad} function of a package

\item \code{roam_activate_all()}: Activate all the roam objects in the given package.
Used in the \link{.onLoad} function of a package

}}
\examples{
# Define the roam object
bee <- new_roam(
  "roam", "bee",
  function(version)
  read.csv(
    "https://raw.githubusercontent.com/finyang/roam/master/demo/bee_colonies.csv"
  ))
# Activation
roam_activate(bee)
if (interactive()) {
  # Download
  roam_install(bee)
  # or in an interative session, simply
  bee
  # Access
  bee
  # Update
  roam_update(bee)
  # Deleting cache
  roam_delete(bee)
}
}
